backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pinecone-client==2.2.4
python-dotenv
pytest
PyPDF2
python-docx
